refactoring add reader

+++ src/main/java/newway/bank/Bank.java
@@ -11,14 +11,14 @@ import java.time.LocalDateTime;
 @EqualsAndHashCode
 @ToString
 public class Bank {
-    private Float USD_buy = 0.0f;
-    private Float USD_sell = 0.0f;
-    private Float EUR_buy = 0.0f;
-    private Float EUR_sell = 0.0f;
-    private Float PLN_buy = 0.0f;
-    private Float PLN_sell = 0.0f;
-    private Float BTC_buy = 0.0f;
-    private Float BTC_sell = 0.0f;
+    private Double USD_buy = 0.0;
+    private Double USD_sell = 0.0;
+    private Double EUR_buy = 0.0;
+    private Double EUR_sell = 0.0;
+    private Double PLN_buy = 0.0;
+    private Double PLN_sell = 0.0;
+    private Double BTC_buy = 0.0;
+    private Double BTC_sell = 0.0;
     private LocalDateTime time;
     private String URL;
 }
